Controllers: - AuthController - Removed login from UserController
